Adds suitable hook scripts for branching and merging
The post-checkout hook will detect if its a branch and if so copy the current development database from the old branch to a new database with the right name for the new branch.
The merge hook will check for each _development and _test database that a corresponding branch still exists locally.  If it does not, then the database is dropped.